requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Excel Processing
openpyxl==3.1.2
xlrd==2.0.1

# Google Drive Integration
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Vector Databases
chromadb>=0.5.0
opensearch-py==2.4.2

# LLM Providers
openai>=1.51.0
anthropic>=0.36.0

# Embedding Providers
sentence-transformers>=3.0.0
cohere==4.47
FlagEmbedding>=1.2.0  # For BGE-M3 (optional, can use sentence-transformers)

# Hybrid Search
rank-bm25>=0.2.2

# Open-Source Document Extraction (optional - all run locally, no API keys needed)
# docling>=1.0.0  # IBM Docling - uncomment to enable
# unstructured[xlsx]>=0.10.0  # Unstructured.io - uncomment to enable

# Local LLM Support (optional)
# ollama>=0.1.0  # Ollama client - uncomment to enable

# Utilities
python-dotenv==1.0.1
cryptography==42.0.2
python-dateutil==2.8.2
dateparser==1.2.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
python-json-logger==2.0.7
psutil==5.9.8
PyJWT==2.8.0
python-multipart==0.0.6

# CLI
click==8.1.7
tqdm==4.66.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
